chore: move definitions to separate files
