+++ src/definitions/index.ts
@@ -1,0 +1,20 @@
+import type { AddressDefinitions } from './address';
+import type { AnimalDefinitions } from './animal';
+import type { Definitions, DefinitionTypes } from './definitions';
+import { definitions } from './definitions';
+import type { NameDefinitions, NameTitleDefinitions } from './name';
+import type { Format, Formats, Texts, Values } from './utils';
+
+export { definitions };
+export type {
+  Definitions,
+  DefinitionTypes,
+  Values,
+  Texts,
+  Format,
+  Formats,
+  AnimalDefinitions,
+  AddressDefinitions,
+  NameDefinitions,
+  NameTitleDefinitions,
+};
